+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Define Field.Set.extend() new method which accepts an array at Field.Set instanciation
     - Define Field.Set.names() new method which returns defined names
     - Field.configure() becomes a reactive data source
+    - Accept dt_data extended definition
 
 ### 1.1.0